requirements.txt
# Core dependencies
streamlit>=1.32.0
typer>=0.9.0
rich>=13.7.0
python-dotenv>=1.0.0
pysrt>=1.1.2

# LLM and AI dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
openai>=1.12.0
google-generativeai>=0.3.2

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Type hints
typing-extensions>=4.9.0

# Progress tracking
tqdm>=4.66.0 